issue 14 multiple fixes
