+++ dispatcher.go
@@ -6,7 +6,6 @@ import (
 	"sync"
 
 	"context"
-	"fmt"
 
 	datastore "github.com/ipfs/go-datastore"
 	"github.com/ipfs/go-datastore/query"
@@ -16,11 +15,6 @@ import (
 type Reducer interface {
 	Reduce(event Event) error
 }
-
-// @todo: Should we also support a `Transformer` to actually fetch raw event data as part of a pipeline?
-
-// Token is a simple unique ID used to reference a registered callback.
-type Token string
 
 const prefix = "ID"
 
@@ -33,15 +27,14 @@ var lastID = 0
 // which can be used to deregister the reducer later.
 type Dispatcher struct {
 	store    datastore.TxnDatastore
-	reducers map[Token]Reducer
+	reducers []Reducer
 	lock     sync.Mutex
 }
 
 // NewDispatcher creates a new EventDispatcher
 func NewDispatcher(store datastore.TxnDatastore) *Dispatcher {
 	return &Dispatcher{
-		store:    store,
-		reducers: make(map[Token]Reducer),
+		store: store,
 	}
 }
 
@@ -50,25 +43,12 @@ func (d *Dispatcher) Store() datastore.TxnDatastore {
 	return d.store
 }
 
-// Register takes a reducer to be invoked with each dispatched event and returns a token for de-registration.
-func (d *Dispatcher) Register(reducer Reducer) Token {
+// Register takes a reducer to be invoked with each dispatched event
+func (d *Dispatcher) Register(reducer Reducer) {
 	d.lock.Lock()
 	defer d.lock.Unlock()
 	lastID++
-	id := Token(fmt.Sprintf("%s-%d", prefix, lastID))
-	d.reducers[id] = reducer
-	return id
-}
-
-// Deregister removes a reducer based on its token.
-func (d *Dispatcher) Deregister(token Token) error {
-	d.lock.Lock()
-	defer d.lock.Unlock()
-	if _, ok := d.reducers[token]; !ok {
-		return fmt.Errorf("`%s` does not map to a registered callback", token)
-	}
-	delete(d.reducers, token)
-	return nil
+	d.reducers = append(d.reducers, reducer)
 }
 
 // Dispatch dispatches a payload to all registered reducers.
