+++ dispatcher_test.go
@@ -19,10 +19,7 @@ func TestNewEventDispatcher(t *testing.T) {
 func TestRegister(t *testing.T) {
 	eventstore := NewTxMapDatastore()
 	dispatcher := NewDispatcher(eventstore)
-	token := dispatcher.Register(&nullReducer{})
-	if token != "ID-1" {
-		t.Error("callback registration failed")
-	}
+	dispatcher.Register(&nullReducer{})
 	if len(dispatcher.reducers) < 1 {
 		t.Error("expected callbacks map to have non-zero length")
 	}
@@ -52,21 +49,6 @@ func TestDispatchLock(t *testing.T) {
 	}
 }
 
-func TestDeregister(t *testing.T) {
-	eventstore := NewTxMapDatastore()
-	dispatcher := NewDispatcher(eventstore)
-	if err := dispatcher.Deregister("string"); err == nil {
-		t.Error("expected invalid de-registration to return error")
-	}
-	token := dispatcher.Register(&nullReducer{})
-	if err := dispatcher.Deregister(token); err != nil {
-		t.Error("error attempting to deregister a valid callback")
-	}
-	if len(dispatcher.reducers) > 0 {
-		t.Error("expected callbacks map to have zero length")
-	}
-}
-
 func TestDispatch(t *testing.T) {
 	eventstore := NewTxMapDatastore()
 	dispatcher := NewDispatcher(eventstore)
@@ -75,6 +57,9 @@ func TestDispatch(t *testing.T) {
 		t.Error("unexpected error in dispatch call")
 	}
 	results, err := dispatcher.Query(query.Query{})
+	if err != nil {
+		t.Fatalf("query failed: %v", err)
+	}
 	if len(results) != 1 {
 		t.Errorf("expected 1 result, got %d", len(results))
 	}
@@ -87,6 +72,9 @@ func TestDispatch(t *testing.T) {
 		t.Errorf("`%s` should be `error`", err)
 	}
 	results, err = dispatcher.Query(query.Query{})
+	if err != nil {
+		t.Fatalf("query failed: %v", err)
+	}
 	if len(results) > 1 {
 		t.Errorf("expected 1 result, got %d", len(results))
 	}
